+++ src/config-schema.ts
@@ -123,9 +123,9 @@ export const configSchema = {
   },
   enableStockDispense: {
     _type: Type.Boolean,
-    _description: 'To enable or disable the deduction of stock during the dispensing process',
-    _default: true,
-  },
+    _description: 'Enable or disable stock deduction during the dispensing process. Requires the stock management module to be installed and configured.',
+    _default: false,
+},
 };
 
 export interface PharmacyConfig {
