+++ src/forms/stock-dispense/stock-dispense.component.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ComboBox, InlineLoading, InlineNotification, Layer } from '@carbon/react';
 import { type MedicationDispense, type InventoryItem } from '../../types';
-import { useDispenseStock } from './useDispenseStock';
+import { useDispenseStock } from './stock.resource';
 import { formatDate } from '@openmrs/esm-framework';
 import { useTranslation } from 'react-i18next';
